move collision check to utilities
